+++ src/SearchInput/SearchInput.stories.tsx
@@ -1,13 +1,18 @@
 import React from 'react'
-import { SearchInput, SearchInputProps } from './SearchInput'
+import { Box } from '../Box'
 import { Container } from './Container'
+import { SearchInput, SearchInputProps } from './SearchInput'
 
 export default {
   title: 'Search Input',
   component: SearchInput,
 }
 
-const Template = (props: SearchInputProps) => <Container {...props} />
+const Template = (props: SearchInputProps) => (
+  <Box height="500px" width="100%">
+    <Container {...props} />
+  </Box>
+)
 
 export const Default = Template.bind({})